# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.ID:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  requests.Pagination:
    properties:
      page:
        minimum: 1
        type: integer
    required:
    - page
    type: object
...
  requests.Search:
    properties:
      page:
        minimum: 1
        type: integer
      search:
        type: string
    required:
    - page
    - search
    type: object
  requests.SortByYearSeasonAnime:
    properties:
      page:
        minimum: 1
        type: integer
      season:
        enum:
        - winter
        - summer
        - fall
        - spring
        type: string
      sort:
        enum:
        - popularity
        - new
        - old
        type: string
      year:
        type: integer
    required:
    - page
    - season
    - sort
    - year
    type: object
  requests.SortFilterAnime:
    properties:
      demographics:
        type: string
      genres:
        type: string
      page:
        minimum: 1
        type: integer
      sort:
        enum:
        - top
        - popularity
        - new
        - old
        type: string
      status:
        enum:
        - airing
        - upcoming
        - finished
        type: string
      studios:
        type: string
      themes:
        type: string
    required:
    - page
    - sort
    type: object
...
  requests.SortFilterGame:
    properties:
      genres:
        type: string
      page:
        minimum: 1
        type: integer
      platform:
        type: string
      sort:
        enum:
        - top
        - popularity
        - new
        - old
        type: string
      tba:
        type: boolean
    required:
    - page
    - sort
    type: object
...
  responses.Anime:
    properties:
      _id:
        type: string
      age_rating:
        type: string
      aired:
        $ref: '#/definitions/responses.AnimeAirDate'
      characters:
        items:
          $ref: '#/definitions/responses.AnimeCharacter'
        type: array
      demographics:
        items:
          $ref: '#/definitions/responses.AnimeGenre'
        type: array
      description:
        type: string
      episodes:
        type: integer
      genres:
        items:
          $ref: '#/definitions/responses.AnimeGenre'
        type: array
      image_url:
        type: string
      is_airing:
        type: boolean
      mal_id:
        type: integer
      mal_score:
        type: number
      mal_scored_by:
        type: integer
      popularity:
        type: number
      producers:
        items:
          $ref: '#/definitions/responses.AnimeNameURL'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/responses.AnimeRecommendation'
        type: array
      relations:
        items:
          $ref: '#/definitions/responses.AnimeRelation'
        type: array
      reviews:
        $ref: '#/definitions/responses.ReviewSummary'
      season:
        type: string
      source:
        type: string
      status:
        type: string
      streaming:
        items:
          $ref: '#/definitions/responses.AnimeNameURL'
        type: array
      studios:
        items:
          $ref: '#/definitions/responses.AnimeNameURL'
        type: array
      themes:
        items:
          $ref: '#/definitions/responses.AnimeGenre'
        type: array
      title_en:
        type: string
      title_jp:
        type: string
      title_original:
        type: string
      trailer:
        type: string
      type:
        type: string
      year:
        type: integer
    type: object
  responses.AnimeAirDate:
    properties:
      from:
        type: string
      from_day:
        type: integer
      from_month:
        type: integer
      from_year:
        type: integer
      to:
        type: string
      to_day:
        type: integer
      to_month:
        type: integer
      to_year:
        type: integer
    type: object
  responses.AnimeCharacter:
    properties:
      image:
        type: string
      mal_id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
  responses.AnimeDetails:
    properties:
      _id:
        type: string
      age_rating:
        type: string
      aired:
        $ref: '#/definitions/responses.AnimeAirDate'
      anime_list:
        $ref: '#/definitions/responses.AnimeWatchList'
      characters:
        items:
          $ref: '#/definitions/responses.AnimeCharacter'
        type: array
      demographics:
        items:
          $ref: '#/definitions/responses.AnimeGenre'
        type: array
      description:
        type: string
      episodes:
        type: integer
      genres:
        items:
          $ref: '#/definitions/responses.AnimeGenre'
        type: array
      image_url:
        type: string
      is_airing:
        type: boolean
      mal_id:
        type: integer
      mal_score:
        type: number
      mal_scored_by:
        type: integer
      producers:
        items:
          $ref: '#/definitions/responses.AnimeNameURL'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/responses.AnimeRecommendation'
        type: array
      relations:
        items:
          $ref: '#/definitions/responses.AnimeRelation'
        type: array
      reviews:
        $ref: '#/definitions/responses.ReviewSummary'
      season:
        type: string
      source:
        type: string
      status:
        type: string
      streaming:
        items:
          $ref: '#/definitions/responses.AnimeNameURL'
        type: array
      studios:
        items:
          $ref: '#/definitions/responses.AnimeNameURL'
        type: array
      themes:
        items:
          $ref: '#/definitions/responses.AnimeGenre'
        type: array
      title_en:
        type: string
      title_jp:
        type: string
      title_original:
        type: string
      trailer:
        type: string
      type:
        type: string
      watch_later:
        $ref: '#/definitions/responses.ConsumeLater'
      year:
        type: integer
    type: object
  responses.AnimeGenre:
    properties:
      name:
        type: string
      url:
        type: string
    type: object
...
  responses.AnimeNameURL:
    properties:
      name:
        type: string
      url:
        type: string
    type: object
  responses.AnimeRecommendation:
    properties:
      image_url:
        type: string
      mal_id:
        type: integer
      title:
        type: string
    type: object
  responses.AnimeRelation:
    properties:
      _id:
        type: string
      anime_id:
        type: string
      image_url:
        type: string
      mal_id:
        type: integer
      relation:
        type: string
      title_en:
        type: string
      title_original:
        type: string
      type:
        type: string
    type: object
  responses.AnimeWatchList:
    properties:
      _id:
        type: string
      anime_id:
        type: string
      anime_mal_id:
        type: integer
      created_at:
        type: string
      score:
        type: number
      status:
        type: string
      times_finished:
        type: integer
      user_id:
        type: string
      watched_episodes:
        type: integer
    type: object
  responses.Author:
    properties:
      _id:
        type: string
      email:
        type: string
      image:
        type: string
      is_premium:
        type: boolean
      username:
        type: string
    type: object
  responses.ConsumeLater:
    properties:
      _id:
        type: string
      content:
        $ref: '#/definitions/responses.ConsumeLaterContent'
      content_external_id:
        type: string
      content_external_int_id:
        type: integer
      content_id:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      self_note:
        type: string
      user_id:
        type: string
    type: object
  responses.ConsumeLaterContent:
    properties:
      description:
        type: string
      image_url:
        type: string
      score:
        type: number
      title_en:
        type: string
      title_original:
        type: string
    type: object
  responses.DayOfWeekAnime:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.Anime'
        type: array
      day_of_week:
        type: integer
    type: object
...
  responses.Game:
    properties:
      _id:
        type: string
      age_rating:
        type: string
      description:
        type: string
      developers:
        items:
          type: string
        type: array
      genres:
        items:
          type: string
        type: array
      image_url:
        type: string
      metacritic_score:
        type: integer
      metacritic_score_by_platform:
        items:
          $ref: '#/definitions/responses.GameMetacriticScorePlatform'
        type: array
      platforms:
        items:
          type: string
        type: array
      popularity:
        type: number
      publishers:
        items:
          type: string
        type: array
      rawg_id:
        type: integer
      rawg_rating:
        type: number
      rawg_rating_count:
        type: integer
      related_games:
        items:
          $ref: '#/definitions/responses.GameRelation'
        type: array
      release_date:
        type: string
      reviews:
        $ref: '#/definitions/responses.ReviewSummary'
      screenshots:
        items:
          type: string
        type: array
      stores:
        items:
          $ref: '#/definitions/responses.GameStore'
        type: array
      tags:
        items:
          type: string
        type: array
      tba:
        type: boolean
      title:
        type: string
      title_original:
        type: string
    type: object
  responses.GameDetails:
    properties:
      _id:
        type: string
      age_rating:
        type: string
      description:
        type: string
      developers:
        items:
          type: string
        type: array
      game_list:
        $ref: '#/definitions/responses.GamePlayList'
      genres:
        items:
          type: string
        type: array
      image_url:
        type: string
      metacritic_score:
        type: integer
      metacritic_score_by_platform:
        items:
          $ref: '#/definitions/responses.GameMetacriticScorePlatform'
        type: array
      platforms:
        items:
          type: string
        type: array
      publishers:
        items:
          type: string
        type: array
      rawg_id:
        type: integer
      rawg_rating:
        type: number
      rawg_rating_count:
        type: integer
      related_games:
        items:
          $ref: '#/definitions/responses.GameDetailsRelation'
        type: array
      release_date:
        type: string
      reviews:
        $ref: '#/definitions/responses.ReviewSummary'
      screenshots:
        items:
          type: string
        type: array
      stores:
        items:
          $ref: '#/definitions/responses.GameStore'
        type: array
      tags:
        items:
          type: string
        type: array
      tba:
        type: boolean
      title:
        type: string
      title_original:
        type: string
      watch_later:
        $ref: '#/definitions/responses.ConsumeLater'
    type: object
  responses.GameDetailsRelation:
    properties:
      _id:
        type: string
      image_url:
        type: string
      platforms:
        items:
          type: string
        type: array
      rawg_id:
        type: integer
      release_date:
        type: string
      title:
        type: string
      title_original:
        type: string
    type: object
...
  responses.GameMetacriticScorePlatform:
    properties:
      platform:
        type: string
      score:
        type: number
    type: object
  responses.GamePlayList:
    properties:
      _id:
        type: string
      achievement_status:
        type: number
      created_at:
        type: string
      game_id:
        type: string
      game_rawg_id:
        type: integer
      hours_played:
        type: integer
      score:
        type: number
      status:
        type: string
      times_finished:
        type: integer
      user_id:
        type: string
    type: object
  responses.GameRelation:
    properties:
      name:
        type: string
      rawg_id:
        type: integer
      release_date:
        type: string
    type: object
  responses.GameStore:
    properties:
      store_id:
        type: integer
      url:
        type: string
    type: object
...
  responses.Review:
    properties:
      _id:
        type: string
      author:
        $ref: '#/definitions/responses.Author'
      content_external_id:
        type: string
      content_external_int_id:
        type: integer
      content_id:
        type: string
      created_at:
        type: string
      is_author:
        type: boolean
      is_liked:
        type: boolean
      likes:
        items:
          type: string
        type: array
      popularity:
        type: integer
      review:
        type: string
      star:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  responses.ReviewSummary:
    properties:
      avg_star:
        type: number
      is_reviewed:
        type: boolean
      review:
        $ref: '#/definitions/responses.Review'
      star_counts:
        $ref: '#/definitions/responses.StarCounts'
      total_votes:
        type: integer
    type: object
...
  responses.StarCounts:
    properties:
      five_star:
        type: integer
      four_star:
        type: integer
      one_star:
        type: integer
      three_star:
        type: integer
      two_star:
        type: integer
    type: object
...
host: http://localhost:8080
info:
  contact:
    email: mrntlu@gmail.com
    name: Burak Fidan
  description: REST Api of Project Consumer.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: '#soon'
  title: Project Consumer API
  version: "1.0"
paths:
  /anime:
    get:
      consumes:
      - application/json
      description: Returns animes by sort and filter
      parameters:
      - description: Sort and Filter Anime
        in: body
        name: sortfilteranime
        required: true
        schema:
          $ref: '#/definitions/requests.SortFilterAnime'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Anime'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Animes by Sort and Filter
      tags:
      - anime
  /anime/airing:
    get:
      consumes:
      - application/json
      description: Returns upcoming animes by day of week
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.DayOfWeekAnime'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Upcoming Animes by Day of Week
      tags:
      - anime
  /anime/details:
    get:
      consumes:
      - application/json
      description: Returns anime details with optional authentication
      parameters:
      - description: ID
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.AnimeDetails'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Anime Details
      tags:
      - anime
  /anime/search:
    get:
      consumes:
      - application/json
      description: Search animes
      parameters:
      - description: Search
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/requests.Search'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Anime'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search Anime
      tags:
      - anime
  /anime/season:
    get:
      consumes:
      - application/json
      description: Returns animes by year and season
      parameters:
      - description: Sort Anime By Year and Season
        in: body
        name: sortbyyearseasonanime
        required: true
        schema:
          $ref: '#/definitions/requests.SortByYearSeasonAnime'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Anime'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Animes by Year and Season
      tags:
      - anime
  /anime/upcoming:
    get:
      consumes:
      - application/json
      description: Returns upcoming animes by sort with pagination
      parameters:
      - description: Pagination
        in: body
        name: pagination
        required: true
        schema:
          $ref: '#/definitions/requests.Pagination'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Anime'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Upcoming Animes by Sort
      tags:
      - anime
...
  /game:
    get:
      consumes:
      - application/json
      description: Returns games by sort and filter
      parameters:
      - description: Sort and Filter Game
        in: body
        name: sortfiltergame
        required: true
        schema:
          $ref: '#/definitions/requests.SortFilterGame'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Game'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Games by Sort and Filter
      tags:
      - game
  /game/details:
    get:
      consumes:
      - application/json
      description: Returns game details with optional authentication
      parameters:
      - description: ID
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.GameDetails'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Game Details
      tags:
      - game
  /game/search:
    get:
      consumes:
      - application/json
      description: Search Game
      parameters:
      - description: ID
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      - description: Search
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/requests.Search'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Game'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search Game
      tags:
      - game
  /game/upcoming:
    get:
      consumes:
      - application/json
      description: Returns upcoming games by sort with pagination
      parameters:
      - description: Pagination
        in: body
        name: pagination
        required: true
        schema:
          $ref: '#/definitions/requests.Pagination'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Game'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Upcoming Games by Sort
      tags:
      - game
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"